Add new library-functions and layout
Added new array and filled it with 6 Books.
Added a displayBooks-function to display all books on the page (not
complete yet).
Added CSS-layouts with header, sidebar an main-content grid-areas.
Added some base style for better appearance.
Added some placeholder-Books to see how the site will look with a filled library.